+++ utils/http.go
@@ -91,3 +91,7 @@ func RequestHttp[TResult any](url string, payload any, headers map[string]string
 
 	return result, nil
 }
+
+func WriteJsonContentHeader(w http.ResponseWriter) {
+	w.Header().Set("Content-Type", "application/json")
+}